getAutomanCallData
